Rename Director to Directord
Sadly the director name was already taken on pypi and there
was no way I could assume the director project name, even
though there's been no work on it pushlished since 2009.
This PR renames the Director project to "Directord" and
changes the install from the git source to using the pypi
published package.

Signed-off-by: Kevin Carter <kecarter@redhat.com>