Added needed-by to tasks
taskflow only allows for provides and requires. But if we want to be
able to dynamically add requirements to other tasks we need to do the
processing in task-core instead. This change adds a needed-by key to
tasks and additional functions that can be used after loading all the
services but before adding services to the taskflow flow to enable
needed-by functionality.