Add task_options to ansible runner tasks
In order to be able to customize extra things, we need a way to pass the
options to the task runner via the task.  This change adds a
task_options dict to the task definition which would be used for this.
The first option that leverages this would be the
ANSIBLE_STDOUT_CALLBACK configuration.